+++ src/main/java/com/rakuten/felix/testsend/manager/webclients/dto/Column.java
@@ -1,0 +1,17 @@
+package com.rakuten.felix.testsend.manager.webclients.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.Builder;
+import lombok.Value;
+
+@Value
+@Builder
+@JsonIgnoreProperties(ignoreUnknown = true)
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class Column {
+    Integer index;
+    Integer attributeIndex;
+    String keyword;
+    String replace;
+}