+++ src/main/java/com/rakuten/felix/testsend/manager/webclients/dto/MailJob.java
@@ -18,4 +18,5 @@ public class MailJob {
     List<String> parts;
     @NotNull
     List<String> prependAddresses;
+    Columns columns;
 }