+++ src/test/java/com/rakuten/felix/testsend/manager/MailContentBuilderTest.java
@@ -30,8 +30,9 @@ class MailContentBuilderTest {
     @Test
     void buildHtmlContent_contentIsMultipart_simpleSplit1_segmentSplit1() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getMultipartContentsWithSplit1Segment1();
-        val result = mailContentBuilder.buildHtmlContents(contents, parts);
+        val result = mailContentBuilder.buildHtmlContents(contents, parts, columns);
         val expected = Collections.singletonList("Part0Part1Part2");
         assertResult(expected, result);
     }
@@ -39,8 +40,9 @@ class MailContentBuilderTest {
     @Test
     void buildHtmlContent_contentIsMultipart_simpleSplit1_segmentSplit2() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getMultipartContentsWithSplit1Segment2();
-        val result = mailContentBuilder.buildHtmlContents(contents, parts);
+        val result = mailContentBuilder.buildHtmlContents(contents, parts, columns);
         val expected = Arrays.asList("Part0Part1", "Part2Part3");
         assertResult(expected, result);
     }
@@ -48,8 +50,9 @@ class MailContentBuilderTest {
     @Test
     void buildHtmlContent_contentIsMultipart_simpleSplit2_segmentSplit1() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getMultipartContentsWithSplit2Segment1();
-        val result = mailContentBuilder.buildHtmlContents(contents, parts);
+        val result = mailContentBuilder.buildHtmlContents(contents, parts, columns);
         val expected = Arrays.asList("Part0Part1", "Part2Part3");
         assertResult(expected, result);
     }
@@ -57,8 +60,9 @@ class MailContentBuilderTest {
     @Test
     void buildHtmlContent_contentIsMultipart_simpleSplit2_segmentSplit2() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getMultipartContentsWithSplit2Segment2();
-        val result = mailContentBuilder.buildHtmlContents(contents, parts);
+        val result = mailContentBuilder.buildHtmlContents(contents, parts, columns);
         val expected = Arrays.asList("Part0", "Part1", "Part2", "Part3");
         assertResult(expected, result);
     }
@@ -66,8 +70,9 @@ class MailContentBuilderTest {
     @Test
     void buildHtmlContent_contentIsText_simpleSplit1_segmentSplit1() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getTextContentsWithSplit1Segment1();
-        val result = mailContentBuilder.buildHtmlContents(contents, parts);
+        val result = mailContentBuilder.buildHtmlContents(contents, parts, columns);
         val expected = new ArrayList<String>();
         assertResult(expected, result);
     }
@@ -75,8 +80,9 @@ class MailContentBuilderTest {
     @Test
     void buildHtmlContent_contentIsText_simpleSplit1_segmentSplit2() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getTextContentsWithSplit1Segment2();
-        val result = mailContentBuilder.buildHtmlContents(contents, parts);
+        val result = mailContentBuilder.buildHtmlContents(contents, parts, columns);
         val expected = new ArrayList<String>();
         assertResult(expected, result);
     }
@@ -84,8 +90,9 @@ class MailContentBuilderTest {
     @Test
     void buildHtmlContent_contentIsText_simpleSplit2_segmentSplit1() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getTextContentsWithSplit2Segment1();
-        val result = mailContentBuilder.buildHtmlContents(contents, parts);
+        val result = mailContentBuilder.buildHtmlContents(contents, parts, columns);
         val expected = new ArrayList<String>();
         assertResult(expected, result);
     }
@@ -93,8 +100,9 @@ class MailContentBuilderTest {
     @Test
     void buildHtmlContent_contentIsText_simpleSplit2_segmentSplit2() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getTextContentsWithSplit2Segment2();
-        val result = mailContentBuilder.buildHtmlContents(contents, parts);
+        val result = mailContentBuilder.buildHtmlContents(contents, parts, columns);
         val expected = new ArrayList<String>();
         assertResult(expected, result);
     }
@@ -102,8 +110,9 @@ class MailContentBuilderTest {
     @Test
     void buildHtmlContent_contentIsHtml_simpleSplit1_segmentSplit1() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getHtmlContentsWithSplit1Segment1();
-        val result = mailContentBuilder.buildHtmlContents(contents, parts);
+        val result = mailContentBuilder.buildHtmlContents(contents, parts, columns);
         val expected = Collections.singletonList("Part0Part1Part2");
         assertResult(expected, result);
     }
@@ -111,8 +120,9 @@ class MailContentBuilderTest {
     @Test
     void buildHtmlContent_contentIsHtml_simpleSplit1_segmentSplit2() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getHtmlContentsWithSplit1Segment2();
-        val result = mailContentBuilder.buildHtmlContents(contents, parts);
+        val result = mailContentBuilder.buildHtmlContents(contents, parts, columns);
         val expected = Arrays.asList("Part0Part1", "Part2Part3");
         assertResult(expected, result);
     }
@@ -120,8 +130,9 @@ class MailContentBuilderTest {
     @Test
     void buildHtmlContent_contentIsHtml_simpleSplit2_segmentSplit1() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getHtmlContentsWithSplit2Segment1();
-        val result = mailContentBuilder.buildHtmlContents(contents, parts);
+        val result = mailContentBuilder.buildHtmlContents(contents, parts, columns);
         val expected = Arrays.asList("Part0Part1", "Part2Part3");
         assertResult(expected, result);
     }
@@ -129,8 +140,9 @@ class MailContentBuilderTest {
     @Test
     void buildHtmlContent_contentIsHtml_simpleSplit2_segmentSplit2() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getHtmlContentsWithSplit2Segment2();
-        val result = mailContentBuilder.buildHtmlContents(contents, parts);
+        val result = mailContentBuilder.buildHtmlContents(contents, parts, columns);
         val expected = Arrays.asList("Part0", "Part1", "Part2", "Part3");
         assertResult(expected, result);
     }
@@ -138,8 +150,9 @@ class MailContentBuilderTest {
     @Test
     void buildTextContent_contentIsMultipart_simpleSplit1_segmentSplit1() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getMultipartContentsWithSplit1Segment1();
-        val result = mailContentBuilder.buildTextContents(contents, parts);
+        val result = mailContentBuilder.buildTextContents(contents, parts, columns);
         val expected = Collections.singletonList("Part3Part4Part5");
         assertResult(expected, result);
     }
@@ -147,8 +160,9 @@ class MailContentBuilderTest {
     @Test
     void buildTextContent_contentIsMultipart_simpleSplit1_segmentSplit2() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getMultipartContentsWithSplit1Segment2();
-        val result = mailContentBuilder.buildTextContents(contents, parts);
+        val result = mailContentBuilder.buildTextContents(contents, parts, columns);
         val expected = Arrays.asList("Part4Part5", "Part6Part7");
         assertResult(expected, result);
     }
@@ -156,8 +170,9 @@ class MailContentBuilderTest {
     @Test
     void buildTextContent_contentIsMultipart_simpleSplit2_segmentSplit1() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getMultipartContentsWithSplit2Segment1();
-        val result = mailContentBuilder.buildTextContents(contents, parts);
+        val result = mailContentBuilder.buildTextContents(contents, parts, columns);
         val expected = Arrays.asList("Part4Part5", "Part6Part7");
         assertResult(expected, result);
     }
@@ -165,8 +180,9 @@ class MailContentBuilderTest {
     @Test
     void buildTextContent_contentIsMultipart_simpleSplit2_segmentSplit2() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getMultipartContentsWithSplit2Segment2();
-        val result = mailContentBuilder.buildTextContents(contents, parts);
+        val result = mailContentBuilder.buildTextContents(contents, parts, columns);
         val expected = Arrays.asList("Part4", "Part5", "Part6", "Part7");
         assertResult(expected, result);
     }
@@ -174,8 +190,9 @@ class MailContentBuilderTest {
     @Test
     void buildTextContent_contentIsText_simpleSplit1_segmentSplit1() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getTextContentsWithSplit1Segment1();
-        val result = mailContentBuilder.buildTextContents(contents, parts);
+        val result = mailContentBuilder.buildTextContents(contents, parts, columns);
         val expected = Collections.singletonList("Part3Part4Part5");
         assertResult(expected, result);
     }
@@ -183,8 +200,9 @@ class MailContentBuilderTest {
     @Test
     void buildTextContent_contentIsText_simpleSplit1_segmentSplit2() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getTextContentsWithSplit1Segment2();
-        val result = mailContentBuilder.buildTextContents(contents, parts);
+        val result = mailContentBuilder.buildTextContents(contents, parts, columns);
         val expected = Arrays.asList("Part4Part5", "Part6Part7");
         assertResult(expected, result);
     }
@@ -192,8 +210,9 @@ class MailContentBuilderTest {
     @Test
     void buildTextContent_contentIsText_simpleSplit2_segmentSplit1() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getTextContentsWithSplit2Segment1();
-        val result = mailContentBuilder.buildTextContents(contents, parts);
+        val result = mailContentBuilder.buildTextContents(contents, parts, columns);
         val expected = Arrays.asList("Part4Part5", "Part6Part7");
         assertResult(expected, result);
     }
@@ -201,8 +220,9 @@ class MailContentBuilderTest {
     @Test
     void buildTextContent_contentIsText_simpleSplit2_segmentSplit2() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getTextContentsWithSplit2Segment2();
-        val result = mailContentBuilder.buildTextContents(contents, parts);
+        val result = mailContentBuilder.buildTextContents(contents, parts, columns);
         val expected = Arrays.asList("Part4", "Part5", "Part6", "Part7");
         assertResult(expected, result);
     }
@@ -210,8 +230,9 @@ class MailContentBuilderTest {
     @Test
     void buildTextContent_contentIsHtml_simpleSplit1_segmentSplit1() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getHtmlContentsWithSplit1Segment1();
-        val result = mailContentBuilder.buildTextContents(contents, parts);
+        val result = mailContentBuilder.buildTextContents(contents, parts, columns);
         val expected = new ArrayList<String>();
         assertResult(expected, result);
     }
@@ -219,8 +240,9 @@ class MailContentBuilderTest {
     @Test
     void buildTextContent_contentIsHtml_simpleSplit1_segmentSplit2() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getHtmlContentsWithSplit1Segment2();
-        val result = mailContentBuilder.buildTextContents(contents, parts);
+        val result = mailContentBuilder.buildTextContents(contents, parts, columns);
         val expected = new ArrayList<String>();
         assertResult(expected, result);
     }
@@ -228,8 +250,9 @@ class MailContentBuilderTest {
     @Test
     void buildTextContent_contentIsHtml_simpleSplit2_segmentSplit1() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getHtmlContentsWithSplit2Segment1();
-        val result = mailContentBuilder.buildTextContents(contents, parts);
+        val result = mailContentBuilder.buildTextContents(contents, parts, columns);
         val expected = new ArrayList<String>();
         assertResult(expected, result);
     }
@@ -237,8 +260,9 @@ class MailContentBuilderTest {
     @Test
     void buildTextContent_contentIsHtml_simpleSplit2_segmentSplit2() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val contents = FakeData.getHtmlContentsWithSplit2Segment2();
-        val result = mailContentBuilder.buildTextContents(contents, parts);
+        val result = mailContentBuilder.buildTextContents(contents, parts, columns);
         val expected = new ArrayList<String>();
         assertResult(expected, result);
     }
@@ -246,18 +270,50 @@ class MailContentBuilderTest {
     @Test
     void buildSubjects_singleCondition() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val subjects = FakeData.getSubjectsWithSingleCondition();
-        val result = mailContentBuilder.buildSubjectContents(subjects, parts);
-        val expected = Collections.singletonList("Part0Part1Part2");
+        val result = mailContentBuilder.buildSubjectContents(subjects, parts, columns);
+        val expected = Arrays.asList("Part0", "Part1", "Part2");
         assertResult(expected, result);
     }
 
     @Test
     void buildSubjects_multiConditions() {
         val parts = FakeData.getParts();
+        val columns = FakeData.getColumns();
         val subjects = FakeData.getSubjectsWithMultiCondition();
-        val result = mailContentBuilder.buildSubjectContents(subjects, parts);
-        val expected = Arrays.asList("Part0Part1Part2", "Part3Part4Part5");
+        val result = mailContentBuilder.buildSubjectContents(subjects, parts, columns);
+        val expected = Arrays.asList("Part0", "Part1", "Part2", "Part3", "Part4", "Part5");
+        assertResult(expected, result);
+    }
+
+    @Test
+    void buildHtmlContent_contentIsHtml_simpleSplit2_segmentSplit2_hasMuAttributes() {
+        val parts = FakeData.getPartsWithMuAttributes();
+        val columns = FakeData.getColumns();
+        val contents = FakeData.getHtmlContentsWithSplit2Segment2();
+        val result = mailContentBuilder.buildHtmlContents(contents, parts, columns);
+        val expected = Arrays.asList("Part0:###_ATTRIBUTE0_###", "Part1:%%email%%", "Part2:%%name%%", "Part3:%%easyId%%%%addA%%");
+        assertResult(expected, result);
+    }
+
+    @Test
+    void buildTextContent_contentIsText_simpleSplit2_segmentSplit2_haMuAttributes() {
+        val parts = FakeData.getPartsWithMuAttributes();
+        val columns = FakeData.getColumns();
+        val contents = FakeData.getTextContentsWithSplit2Segment2();
+        val result = mailContentBuilder.buildTextContents(contents, parts, columns);
+        val expected = Arrays.asList("Part4:%%userId%%,%%addB%%", "Part5:%%addA%%", "Part6:%%cms:test-attrA%%", "Part7:###_ATTRI%%email%%BUTE22_###");
+        assertResult(expected, result);
+    }
+
+    @Test
+    void buildSubjects_multiConditionsAndHasMuAttributes() {
+        val parts = FakeData.getPartsWithMuAttributes();
+        val columns = FakeData.getColumns();
+        val subjects = FakeData.getSubjectsWithMultiCondition();
+        val result = mailContentBuilder.buildSubjectContents(subjects, parts, columns);
+        val expected = Arrays.asList("Part0:###_ATTRIBUTE0_###", "Part1:%%email%%", "Part2:%%name%%", "Part3:%%easyId%%%%addA%%", "Part4:%%userId%%,%%addB%%", "Part5:%%addA%%");
         assertResult(expected, result);
     }
 }